User stats endpoint added

+++ src/main/java/com/multi/shop/api/multi_shop_api/common/controllers/AppController.java
@@ -28,8 +28,8 @@ public class AppController {
 
     @GetMapping("/quantity")
     @PreAuthorize("hasRole('ADMIN')")
-    public Map<String, Integer> appStats() {
-        Map<String, Integer> appStats = new HashMap<>();
+    public Map<String, Long> appStats() {
+        Map<String, Long> appStats = new HashMap<>();
 
         appStats.put("users", userService.usersSize());
         appStats.put("products", productService.productsSize());
