+++ src/main/java/com/multi/shop/api/multi_shop_api/products/services/ProductCategoryService.java
@@ -22,5 +22,5 @@ public interface ProductCategoryService {
 
     List<ProductCategory> findCategoriesByName(List<String> categoryNames);
 
-    int categoriesSize();
+    Long categoriesSize();
 }
