+++ src/main/java/com/multi/shop/api/multi_shop_api/products/services/ProductService.java
@@ -20,5 +20,5 @@ public interface ProductService {
 
     Optional<Product> delete(String id);
 
-    int productsSize();
+    Long productsSize();
 }
